rewrite AsyncPublisherAutoRetryTest
