+++ hermes-test-helper/src/main/java/pl/allegro/tech/hermes/test/helper/containers/ImageTags.java
@@ -2,6 +2,6 @@ package pl.allegro.tech.hermes.test.helper.containers;
 
 public class ImageTags {
     static String confluentImagesTag() {
-        return System.getProperty("confluentImagesTag", "6.1.0");
+        return System.getProperty("confluentImagesTag", "7.2.2");
     }
 }
